+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 - Download pausing and removal to be via uuid instead of index to be more reliable
 - Special pages to no longer use a reversed rendering order, making selecting and styling easier
 - Version and newtab page to no longer use a fixed minimum width
+- Additional cut/copy/paste keys to also work in the navbar (S-del, C-Insert & S-Insert)
 
 ### Fixed